rendering first geojson data
Two new examples VGL2 and VGL3. These extract the coordinates from
a geojson string and render it on the VGL map.

Neither grap the connectivity and draw lines or polys yet, but rather
simply grep coordinates. Second example calls a VTK pipeline to
generate the data first, which requires my addition to VTK for the
GeoJSON writer.