brought back solid color shader and cleaned up detection of scalar type
Now color type choice is based on the presence of the scalar field, or color field, and lookup table range is automatic.
Brought back the solid color shaders, as we need a way to specify the one color in the case when the data itself lacks colors (e.g. country lines).
Added this as a new material type, so that it doesn't preclude the use of the old material which assumes color is present. More cleanup needs to be done here, the solid color should be specified on the "actor" level, like opacity is done now.