Allow geoweb to be run from source tree
Allow geoweb to be run from source tree to speed up development
cycle. Extend geoweb to allow a path to be passed in so the path
to the source tree can be used to serve static content such as
HTML and JavaScript. Create web/lib/geoweb.min.js that lives in
the source tree that dynamically loads the required JavaScript
file from the source tree.

The DEV_BUILD option enables this configuration. This will
create a geoweb_dev script in the deploy directory that
can be used to start geoweb in the source tree.