Graph features detached from the scene tree
The graph feature is a composite feature containing points and
lines, which should not be added to the layer's scene tree.  This
commit adds the ability to create a detached feature for this
purpose.  Because the graph manages its own children, events
propagate down from the graph feature automatically, and
there is not duplication due to them being attached to the
layer scene object as well.