Speed up rendering when some points are hidden.
If some points are hidden because they have neither stroke nor fill, or because they have no opacity or size, then set gl_Position to outside of the view to skip the bulk of the computation for it.

Use an equilateral triangle surrounding a point rather than a square.  This reduces the number of vertices.

Have the fragment shader terminate when outside of the radius.