Added dynamic drawing capability to line features.
Added a verticesPerFeature call.

Refactored the code which generated the pair of triangles per line segment to use an array of information.  This allows a caller to determine which vertices of the triangles belong to which end of the line.  I expect that I will need this to dynamically change one end of a line without changing the other.