Don't create needless arrays.
We create arrays using the vgl.DataBuffers.  Except for positions, which need to be transformed, load this data buffers directly.  Prior to this, the data was loaded into javascript arrays, and then that was copied to the data buffers.

Fill buffers much more efficiently by doing it directly instead of calling buffers.write and buffers.repeat

Move code from createGLPoints to _init and from createGLLines to _init.  Don't do things more than once if they can be done just once.

The point vertexShader wasn't explicitly asking for highp floats.

Added an early-exit clause to convertColor to speed up handling colors that are already in the preferred format.

Update the reference to vgl.