Do less work to convert constant colors.
When we access a style, we always return a function.  Colors are passed through the util.convertColor function.  This was always wrapped util.covertColor(util.ensureFunction(color)).  However, if the color was not a function, it is more efficient to use util.ensureFunction(util.covertColor(color)), as we know the conversion only needs to be performed once.