Add a wrapLongitude option to the contourFeature.
I am defaulting this to true, such that if stride positioning is used (x0, y0, dx, dy), if the data spans 180 degrees (plus or minus), it will be wrapped around so that part of the contour is at negative longitude and part is at positive longitude.  If the grid appears to span 360 degrees, join the ends (but only if wrapping was necessary).

In the general case, we might want to check the map coordinate system and only default to true if the x coordinate system is degree based (or even adjust it to the coordinate system so you could use gradians).

Added a jasmine test to make sure wrapping behaves as expected.