Fixed two issues with the tile cache.
The _atime array sometimes stored the hash value for a tile and sometimes stored the tile itself.  Always use the hash value.  Additionally, the code that moved a tile to the front of the _atime list ended up storing the tile in an array.

When a tile had fallen out of cache, a new identical tile could be requested.  This could result in two copies of the same tile being in use.  If a tile has fallen out of cache, don't draw it.  Perhaps we should add a waring that the tile cache could be increased (but this might just be the result of rapid navigation and indicates we don't care about that tile anymore).

I added a 'tile-reference' attribute to the tile containers.  This makes it possible to use css to help debug tiles.  For instance, adding

```
.geo-tile-container:after {
    content: attr(tile-reference);
    position: absolute;
    top: 5px;
    left: 10px;
    font-size: 30px;
    color: black;
}
```

to a css file will show the tile hash on each tile.  Adding

```
.geo-tile-container img {
    box-sizing: inherit;
    border: 4px solid rgba(0,0,0,0.25);
}
```

Draws a border on each tile.  This facilitates tile debugging.