Don't draw tiles we no longer are interested in.
If, by the tile a tile's image is loaded, we no longer want to show that tile, don't draw it.  This would happen, for instance, when you zoom in, don't wait until the images are fully loaded, and then zoom out.  Sometimes you would see the zoomed in tiles on top of the current layer; these would vanish as soon as the map was updated for any reason.