Add a d3 tile layer.  Add tile url templates.
This introduces a number of changes that might be better done another way.

Currently, we subclass features, but we don't have an easy way to subclass layer types because of how they are attached to renderers.  Instead, I adjust an existing layer type by taking the instantiated layer and modifying it -- I don't like this, but I haven't decided how it ought to be done.

For d3 layers, I've added the ability to have sub-groups, which allows the tiles to layer properly and gives the multi-level features that we have the in the html renderer.

Some of the layer transitions are tracked by storing them on a item called 'canvasElement'.  'canvas' and 'canvasElement' are probably bad terms, as there is no canvas in the html or d3 renderers, and the distinction is whether we transform an element directly or indirectly.

Otherwise, using d3 versus html seems equally performant in my mild tests, even though html uses img elements in divs and d3 uses svg images in g elements.

The url templating code from PR #463 has been conceptually added (note that what had been 'baseUrl' had been changed earlier in this PR to 'url').