Make the new tile layer work with vgl.
Fixed a number of typos from previous commits.

Added an option to specify the zoom level used in toLocal and fromLocal in the tileLayer calculation.

In a previous commit, I introduced an extra element for tracking canvas transforms (to help the d3 renderer).  This seems extra awkward with vgl, and will be refactored.

The vgl renderer has become mostly simpler, as the camera no longer needs to be maintained except for sending the view and projection matrices over to the vgl camera.  The current method relies of reversing coordinate systems and adding a z-offset.  Only z coordinates between [0, 1] are drawn.  It seems like this would be better if it was [-1, 1] so that altitudes could be more straightforward.

The tile alignment code to get pixel-perfect views at integer zoom levels needs to be reattached.  This can be applied to the html and d3 renderers, too.

This requires the update to vgl.