Add a function to allow changing the tile url easily.
Update the jquery-plugin to use the current tile templating methods and functions.

When zooming a d3 layer, points would jump around.

Whenever we needed to recalculate attributes (such as the stroke width) because of resizing, the locations would be recalculated as well.  These are stored in d3 adjusted screen coordinates (not actual screen coordinates), so the display values must be altered to match the transform applied to the d3 layer.