Purge non-current tile levels when done loading.
If only one level of tiles should be shown, purge the non-current levels when done loading.  This is necessary to properly show semi-transparent tiles.

There is a curious behavior that if a tile fails to load, if the lower level tile is available if will remain shown.  This can be viewed as either a feature (as low resolution information is better than no information) or as a defect (as there may be some upper level tiles on top of the lower level tiles).