Use native arrays rather than vec3 and mat4.
vec3 and mat4 create Float32Array data types by default.  These are lower precision than native arrays, and should only be used if the loss of precision is acceptable or the values must be passed to a WebGL context.

Rather than use vec3.fromValues, in most cases a plain array can be used.  I've added two new utility functions to create vec3 and mat4 compatible arrays that are just native arrays.

If the speed/memory benefit of an explicit array type is beneficial, these could use Float64Array instead (we would have to test if this is actually faster).

For the full benefit of this, the vgl branch of view-array-type is required.  Without that, it will still have some benefit.