Implemented a tile fetch queue.
This is actually a general deferred object fetch queue.  It expects that no more than some number of Deferred objects should be processed at once.  For images in modern browsers, there is standard limit of six images per domain at a time.  This is the default for the fetch queue.  If more items are requested, they are further deferred and are loaded in a priority order of last-requested-first.  Items can be cancelled as well if they are no longer needed.

Also fixed the metrics that determine the load order of tiles.  There was an issue that tiles were passed and the function didn't know how to access tile.index.  This metric has been extended to have a concept of a 'bottomLevel', where anything lower than that level is lower priority.  This allows highly zoomed areas to show useful tiles first, rather than showing the absolute lowest tiles first.