Reduce the number of times _update is called on the tile layer.
The layer and featureLayer inits were both binding geo.event.*, resulting in two calls to _update each time an event was triggered.  Add a flag to the layer init function to avoid binding events when we know we are going to bind them in the subclass.

We ALWAYS fire a pan event after a zoom event when we generate the events internally.  Therefore, on the tileLayer, ignore zoom events for _update.  It just wastes time.

On the map object, when we generate events, we tag them with eventType, even though the event code will automatically tag them with event.