Check if a renderer is available.
If not, optionally use a fallback.

This is largely intended to fallback when webGL is unavailable, but it also has the feature that d3 is no longer a hard requirement.

Add a simple phantom test for the jquery-plugin.

Also, added a reload function for the jquery-plugin, mainly to facilitate testing.  This could also help lazy-loading in some instances.