Support map rotations.
Maps now have a rotation property (default=0).  The tiles that are fetched are the covering rectangle of tiles.  This could be optimized so that only tiles that have a portion on the display would be fetched (though calculating that would be more costly).  Rotations work on all renderers.

Added rotation interactions.  By default ctrl+left mouse button will rotation around the center of the map by the angle that the mouse cursor moves around the center of the map.  Ctrl+wheel will rotate in 5 degree increments (strictly, one zoom level wheel movement is the same amount of wheel movement that rotates 6 degrees).

Adjust bounding box to handle clamping when rotated.  This will always let you see the entire map or image, but isn't as intuitive as it could be from a UI standpoint.  Perhaps this should be refactored in a later PR.

Fixed issues with map.bounds and map.zoomAndCenterFromBounds not returning data in the specified gcs (but always in the map gcs).

Tiles in the gl renderer will be pixel-perfect at integer zoom levels for rotations that are multiples of 90 degrees.

Removed the minX, minY, maxX, maxY parameters from the tileLayer and osmLayer classes, as they were not used.

Stopped examples from trying to load a favicon (this can be changed if we ever have a favicon, I was just tired of the 404 errors on it).

Stop allowing text selections in the examples.  When dragging maps, selecting text looks bad.

Added a "Spin to Budapest" option to the transitions example.

Updated the tiles example to allow setting the maxBounds and unitsPerPixel in the query arguments.  Added allowRotation to both the controls and the query arguments.