Added some tests for the new code.
Add a test that compares the null renderer with the d3 renderer.  Due to some wackiness in the phantom svg renderer, this only works at 0 and 180 degree rotations.

Fixed some css issues if the browser has odd line-height and font-size behavior (this affected the null renderer's tile sizes in phantom).

Changed how debug borders are drawn in the null renderer on the tiles example.

Added more coverage for the tileLayer code.