Added tests for map.js.
Removed vestigial references to map.toggle(), as it couldn't have ever worked, and any caller with enough information to use it could have toggle a layer's visibility on their own.

For transitions, deeply copy the options so that we aren't altering a data structure the caller may still be using.

Renamed defaultOpts to opts in a lot of the transition code, as it references the specific options and not the defaults.

Actually attach the geo.event.cancelAnimation and geo.event.cancelNavigation functionality in transitions.