Add some support for altering tile projections.
This is not really sufficient, as reprojecting tiles cause the tiles to have a non-uniform density, which means our calculation of the appropriate tiles to show at different zoom levels is incorrect.  If a projection cannot wrap and wrapping is turned on, we may try to ask for a functionally infinite set of tiles.

Do a little less work while checking the tile queue.