Submit notes when we run a build.
This submits some performance tracking data for the osmLayer as a build note.  This note appears in cdash alongside of the build.

This also exposes a test in the transitions demo (add ?test=true[&renderer=(vgl|canvas|d3|html)] to the url).  This currently only logs test results to the console.  My intent is to show them in a dialog, but I haven't written that yet.

Change the jasmine test time from 5 to 15 seconds.  It would probably be better to make this configurable per test.

Changed the travis build to use -j 3 to parallelize some of the tests.