A few minor changes.
When we exited the vglRenderer, we did not destroy the canvas object.  This means that if you create and destroy vgl layers repeatedly, eventually you will exhaust a browser's available gl contexts (depending on the hardware this could be as few as four gl layers or as many as 16).  Additionally, more cleanup is done to ensure that memory can be released in a timely fashion.

Also, in testing this by toggling between transforms, I've added a guard against invalid projections, so that a error is sent to the console and the state can be recovered.  Before, if a bad projection was set, it was often impossible to reset it to a valid value.

Lastly, when the tile layer is a minimum level that isn't 0 and we want to keep lower levels and we aren't wrapping, then we should keep the entire lower level.  This has two benefits: reprojections can look decent just by setting a minimum level even if their bounds are odd, and padding with a non-zero minimum level still has something to show.  This has one drawback: if you have a very high minimum level (such as 5 or above), too many tiles need to be kept in memory.