Add reprojection example.
The projections are taken from spatialreference.org, with some that look best listed at the top.  The capital cities of the world are taken from geonames.org.

There is a performance issue when a select box exists with a few thousand items, so, by default, only a few dozen projections are shown.  The example can show all of the known projections by commenting out a line.

Fixed a bug in d3 points where the stroke would not appear.  This depends on the order of the style keys in the style dictionary -- if stroke came after strokeColor, the stroke might not appear.

Fixed an issue in vgl points where if a transform added a z-value, they might not be rendered.  This could be changed if we ever address issue #394.

Allow changing a widget's geographic position.

Added the ability to set the maximum bounds on a map after it is created.