Fix the widgets example.
If the widget uses a position with x and y parameters, these are expected to be in the map interface gcs (latitude and longitude).  Rather, one should use top, left, right, and bottom for fixed screen locations.

Also, I noticed of Firefox that the examples had a needless vertical scrollbar, which I have hidden.

This fixes issue #519.