Fix a few issues with transitions.
There were several bugs with momentum, transitions, and spring-back.

If momentum was occurring, a mouse wheel event would take place in the wrong location.  For instance, pan a map such that it has momentum and use the mouse wheel to zoom.  The map zooms, but is pans back suddenly to the start of the momentum location.  Now, when a new mouse action occurs, momentum is canceled.

If a transition was occurring, it would "fight" with mouse interaction.  Now, when a mouse action occurs, a transition is canceled and the transitioncancel event is triggered.

Springback used hard-coded boundaries in a fixed coordinate system.  Now it uses maxBounds in the map coordinate system.

Moved the mockAnimationFrame functions to the testUtils.js file so it can be used in multiple tests.

Added a mockDate routine to facilitate testing momentum and springback.