Allow overriding the tilesAtZoom function.
Prior to this, it was assumed that there was 2^level tiles in each direction at each level.  For non-square tile source (such as slide images), this is not the case.  Unless (until?) the tilesLayer knows how many pixels of tiles are available for the maximum level, the `tilesAtZoom` function needs to change.  This allows it to be specified as an option.

This is related to issue #544, but does not complete that.