Add controls to the heatmap example.
Move the example data set to data.kitware.com.

Allow selecting the cities and earthquake data sets.

Fix an issue with updating the radius or blurRadius on the fly (the circle wasn't being recomputed).

Fix an issue with changing the color gradient on the fly (the gradient wasn't being recomputed).