Add a d3 quadFeature.
This will render nearly identically to the gl quadFeature, with the exception that when an image quad is not a parallelogram.

When the quad is not a parallelogram, the gl renderer will skew the image using two different transforms making the quad contain the full image, but with a transform discontinuity along one of the diagonals.  The d3 render will skew the image with a single transform, and part of the image will not be visible.  For not convex quads, the d3 renderer could display part of the image multiple times.