Improve polygons.
Add a polygon example.

Speed up gl polygons a large amount.  Before, the set of polygons in the example was taking between 240 and 350 ms to regenerate.  The variability was largely caused by garbage collection.  There were many slow points in our processing (earcut only accounts for 35 ms of this time).  Now, much less memory is allocated and deallocated, and some maps have been unrolled to plain functions.  The full generation time is around 60 to 75 ms, with much lower memory turn-over.  Further, if only the style changes, the geometry doesn't need to be recomputed, so just recalculating styles takes 11 ms or so.

Sped up the pointInPolygon function, dropping its time (using the example data) from 10 ms to 2 ms or so.

When mousing over a polygon, these changes mean that it feels 'real-time' instead of very laggy.

Polygons had to be defined as {outer: [(points)], inner: [[(points)], ...]}, with inner being optional.  Now, they can also be defined as [(points)], skipping the container structure, which is the same as {outer: [(points)]}.

Improved the documentation on polygonFeature.