Add stroking to polygons.
This is done by creating a sub-feature of lines for the polygons if the polygon feature has any value other than false for the stroke style (either true or a function).  Lines are created for each exterior and interior loop in the polygons, and are not set for interaction.  Lines are created with the same starting and ending point.

Strokes can have all of the properties that lines get -- strokeWidth, strokeColor, strokeOpacity, strokeStyle.  Lines are rendered with the same renderer as the polygons (and on the same level).

Polygons now have a flag to turn off filling the polygon (though the polygon is still created and still earcut, so this doesn't speed things up for polygons with fill = false).

It would be better to add a flag to the line feature to indicate a closed loop.

Changed the gl.object class to not re-subclass sceneObject, as all gl features are already subclasses of sceneObject.

This requires a vgl update to allow unbinding vertex data.  That change requires an update to the quad feature.

Fixed a typo in the lineFeature.