Fixed inconsistency in pointSearch.
BREAKING CHANGE: The pointSearch function returned different keys depending on whether data was present or not in both the pointFeature and the lineFeature.  Specifically, when there is no data in a feature, or other pointSearch functions (in quad, polygon, and the base class feature), pointSearch returns {index: [...], found: [...]}.  When data is present, the point and line features returned {index: [...], data: [...]}.  This has been changed to always return {index: [...], found: [...]}.

Added unit tests for line features.