Fix an interactor bug with momentum.
Using the default map interactor options, if you right click or right drag on the map, mousemove event cease getting generated.  The action state was marked as zoom or momentum after the mouse up (since it was performing that action), and this is never cleared as clearing is done on mouse up.  This blocked generating mousemove events.

Also, tweak one of the polygon tests to try to eliminate occasional test failures.