Fix an issue with mouse movement during momentum.
We were calculating momentum using some of the mouse properties, even after the mouse was released.  This results is odd behavior if you pan quickly, release the mouse button, then move the mouse in a non-linear way.  The solution is to use a separate momentum record instead of modifying the mouse record for momentum.  This bug was only exposed after fixing that we weren't reporting mouse movement during momentum.