Escape cancels drawing an annotation.
Added the Mousetrap library to handle keyboard actions.

Rather than binding key events to the global document, they are bound to the map node.  This necessitates the map node having a tab index so that it will receive key events.  A tab index is added if not present, but only if a map interactor is specified.  In the event that no interactor is desired, the map can be constructed will a null interactor, rather than creating and then clearing the interactor.