Improve tests not directly related to annotations.
Test short hex color strings.  Test bad polygons.  Test one-vertex lines.

In general tests, if beforeEach and afterEach are at the global score of the test, they apply to ALL tests, not just the describe functions in that file.  This results in odd behavior that shouldn't be relied on.  Fix this in the tests for d3GraphFeature, d3PointFeature, geojsonReader, and heatmap.

In resizing a map, asking for a zero size causes errors (this could happen when the DOM was created without a size and then later resized).  Guard against this in the map, and assign sizes to test divs *before* adding them to the body.

The divs in the mapInteractor tests were relying on the size being affected by the beforeEach and afterEach of other modules, so set those explicitly and adjust the tests accordingly.

Fixed a type on the test-utils when unmocking the VGL renderer.