This is the first step to a positionable slider widget.
It would be nicer if you could change the orientation and which ends had the + and - (or even if they appeared).

You can invoke the slide with the position option, which can contain offsets from the top, left, right, or bottom (or some combination thereof).  It can also take a width and height, which are honored, though height needs to be more than 3.5 times the width or stupid things will happen.