Added a maxIndex function.
This returns the maximum value from the pixelmap index (not the number of distinct indices).  ((maxIndex() + 1) is the useful length of the data array.

Also, don't reload the index values from the pixelmap if it isn't necessary.  When only the data updates, this saves drawing the pixelmap to a canvas.