Improve the speed of pixelmap's update.
Significantly increase the speed by caching more and updating less.  Specifically, the working canvas is maintained between updates, and only changed portions are updated.  canvas.toBlob is used in preference to canvas.toDataURL if it is available.  maxIndex is cached.

Added a fallback for boxSearch to the feature class to suppress some console warnings.