More speed improvements for updates.
Use canvas properties for faster drawing of the final image.

Compute only the extent of the changed area and only update that extent.  This computes colors in one pass and updated pixels in a second pass rather than computing them in the same loop.  The first render is probably very slightly slower.