Allow point annotations to scale with zoom.
Added a scaled property.  If false, points are a fixed size.  If true, the first time the point annotation is rendered, it is changed to the current zoom level.  If a zoom level, the radius of the point is scaled with the map so that it is the specified radius value at the scaled property's zoom level.