Add pixel coordinate utility function.
Add a utility function to make it easier to create maps and tile layers in pixel coordinates with common settings and functions.  It seems wasteful to have this code in each non-map example.

Fix the deepzoom example.  The IIP Server sends partial tiles at the right and bottom edge of the tile set.  Our canvas renderer happens to work with this because it doesn't scale the tile images (which means it won't work on a retina tile).  Added a capabilities flag so we can ask for that feature rather than specifying a renderer.