Fix some feature issues.
Features now have a default boxSearch so that brushing events don't throw errors.

When a mousemove event is processed, less work is done if the point search has no hits.

Mouse buttonDown properties are propagated through to the feature mouseclick event so that the button that was used is known.

Better error handling on feature creation.  This allows the annotationLayer to avoid using a try catch block.

featureLayer's method for replacing the list of features didn't properly add or remove children.  Now, features can be added and removed independently of their creation and destruction, though removing a feature from a layer will still leave it attached to its renderer.

Unit tests for feature and featureLayer.

Fixed some typos in comments.