Replace rather than modify default arrays.
When constructing a mapInteractor with options, make sure the whole set is replaced.  We use jquery's extend method.  When using this on something like $.extend([1, 2, 3], [4, 5]), you end up with [4, 5, 3], but for this purpose we want to end up with just the new array.

Also, give default names and an owner to the mapInteractor map actions so that they can be more easily referenced.