Allow colors to include an alpha value.
Also, handle more css color specifications.  Currently, nothing takes advantage of a color which has an alpha value (though the pixelmap soon will).  It would be nice to be able to use that in addition to opacity styles.

Before, we handled colors of the form rgb object, #rrggbb, #rgb, and css color names.  This adds parsing of #rrggbbaa, #rgba, rgb(), rgba(), hsl(), hsla(), and 'transparent', conforming that parsing to the css working group's current draft standard (there are other color formats in that draft that are not implemented).  It also adds one additional css color name, as per that draft.