Some css fixes.
If you include Bootstrap, it changes canvas element's css to be display: inline-block, which results in a peculiar overflow on these elements, showing scroll bars on a map that occupies the full viewport.  By setting canvas elements to display: block, this prevents this problem.

This change makes it so that overflow: hidden is no longer needed on the examples.

Also, on the examples, we were using a relative position of -21 pixels to subtract out the navbar's bottom margin.  Rather than do this, just remove the navbar's bottom margin.

The heatmap exa,ple thumbnail was a different aspect ratio than all of the other example thumbnails.